// src/app/modules/complianceEthics/compliance.interface.ts
export type TCompliance = {
  images: string[];
  description: string;
  social_compliance_img: string[];
  social_compliance_title: string;
  social_compliance_description: string;
  EmployeesCocCovers: string[];
  zeroTolerance_img: string[];
  zeroTolerance_Title: string;
  zeroTolerance_description: string;
  cod_of_conduct_img: string[];
  cod_of_conduct_title: string;
  cod_of_conduct_short_description: string;
  cod_of_conduct_description: string;
  csr_description: string;
  csr_short_description: string;
  csr_title: string;
  csr_img: string[];
};